Move exception to exceptions module
